Major refactor/cleanup and UI improvements.
Added footer for modifiers. Updated UI for header and main puzzle area.
Cleaned up and refactored item generation. Refactored terminus UI so
that the result is a single shape with a transparent cutout instead of a
layered shape that was not see-through. This will make it much easier to
layer all of the items properly (especially beams).

Beams are basically disabled right now, they need to be refactored next.
Modifiers are only placeholders for now, they also need to be wired up.