Move modifiers from item to tile
This allows the modifier toolbar to function properly when there are
multiple items in a tile. The modifiers will now apply to any item in
the tile that support the modifier type.

Terminus UI is also being updated (work in progress). The wall for the
terminus has been removed in favor of just using the wall item in
conjunction with a terminus.