First pass at improving beam step logic
Each step is now the size of half a tile (inradius). Since beams
originate in the center of a tile, this means they will now reach each
other at the same time if pointing in opposite directions. This makes
more sense visually.

Some re-organization was done to handling beam collisions by item type.
This probably needs more work, it feels a bit janky.

There are still bugs in keeping beam items in sync with tiles that will
need to be fixed tomorrow.