Re-do beam updateState functionality
Previously, when the state of a beam needed to be updated, it would chop
off the beam and underlying steps before attempting to re-build from
there. Now it's a little smarter, it will start re-building and only
update the visuals of the beam when necessary.

There are still bugs with the onEvent handling that cause the beam to
not render correctly, but I have to get to a stopping point so it's
gonna be broken for now.