First pass at beam reflections
Oof. Dealing with all the differing directions is getting to be a pain.
Need to refactor them into one unified system at some point.

There's still some bugs when dealing with reflector collisions when
re-evaluating history. There's probably a better way to deal with the
reflections and the fact that the beam collides with the reflector twice
(once going into it, and once going out of it).