Fix drawing of beams in relation to other items
The previous implementation was resulting in beam path items existing
outside of the beam group, so they were not being cleaned up properly
when history was updated. It seems that in PaperJS a group has it's own
hierarchy in relation to other groups/layers. Because beams will need to
be drawn both above and below other items, we can't use a single group
to encompass all of the path items for the beam. So instead of using a
group, we will just add path items for the beam to the items layer, and
add/remove them from the scene manually.

This patch also updates portals to support no defined direction. Portals
with no defined direction can accept any number of beams which will
continue traveling in the direction they were travling upon entering the
portal when they exit. A new puzzle was added to test this. It is not
yet complete, as merging of beams which are travling in the same
direction upon exiting a portal has not been implemented yet.

This patch also includes updates that fix some bugs with the hit testing
of tiles, and it also adds data to each item group so that the group can
be easily identified.