Improved collision handling
Portals are still a bit buggy and will need further iteration, but
collisions are now being sorted by item sortOrder and then by proximity
of the collision to the target point. This allowed beam collision
tracking to be moved back into beam.onCollision and out of
portal.onCollision.

The remaining bugs are mostly with complex interactions between serveral
beams and portal entrances/exits.