Lots of bug fixes for beam collisions and updating
Most of these are related to portals, but not all of them. There is
still one existing bug I'm aware of that has not been fixed, and that's
a collision with a beam in an exit portal, where there is only a single
point for the beam so it doesn't get picked up by
path.getIntersections(). Will have to fix that one in a subsequent patch
as this one is getting a bit large.