Fix merged beam collision loop and move mask
Fixed a beam collision loop on portal exit due to the previously added
fix for colliding with isolated beam path items. The collision will
still occur, but will be ignored if the beam being collided with has
already merged into the beam that is trying to step.

Also fixed a bug with the move modifier mask. Beams now update when
removed, which was causing the puzzle to be solved prior to the move
modifier mask being removed, so it is now removed prior to any items
being moved.