Fix step tile selector logic; update UI
Previously the tile for a step was queried for using the point of the
next step. However, since every step is approximately half a tile in
length, every other step was right on the boundary between two tiles.
This was causing the tile associated with that step to be
undeterministic depending on the exact rounding of the step point. I've
updated the logic to instead use the mid-point between the previous
and next steps, which will always deterministically choose the same
tile. It also makes the most sense intuitively, since that step largely
took place in the tile the mid-point resides in.

The UI has also been updated in preperation for undo/redo functionality.
The previous and next puzzle selectors, as well as reset, are now always
visible and reside next to the puzzle selection dropdown.