Add state storage for toggle and rotate
Also simplified the puzzle configuration syntax to remove the
unnecessary layout.tiles object. There were a lot of other refactorings
in here in order to properly support correct rotation of items based on
the rotation direction stored in state as well as being able to move
that logic into the rotatable class instead of having to put it into
every class which implemented that class.

It's become apparent that it would be much easier to use something like
JSONPath to target pieces of the state to update rather than using row,
column, index and key, which is pretty inflexible. That will be in the
next update.