Finish initial state implementation
This includes support for moving, rotating and toggling items. One thing
not currently supported is anything related to beams; for example, which
portal destinations have been chosen. This state could likely be added
to the existing beam configuration in the initial state of the puzzle
somehow, need to think about it a bit.

The current implementation can also be brittle since it uses array index
positions when updating items inside arrays. I don't foresee any issues
with this, but it could be problematic.