Switch to JSON diff for state changes
On every state change, the current state is now re-generated, a delta
between previous and current state is created and stored in the state
cache. Every object is now responsible for maintaining its own cache.

This contains a fair amount of refactoring to introduce a Stateful
interface for objects as well as avoiding polluting the initial state
with configuration changes, which have been separated out into separate
classes and objects.

In addition to the updates to state, this commit also contains a bug fix
for out of bounds collisions not showing up and escaping of the
generated base64 that gets stored in the URL (to avoid having to encode
any special characters which make the URL longer).