URL and UI improvements
URL now uses path segments instead of parameters for ID and state. The
puzzle selector has been moved to the center of the header and now
displays puzzle titles instead of IDs. The power icon has been replaced
with a link icon (better representation of 'connection'). The message in
the center of the footer no longer displays puzzle title and instead can
optionally display a message from the puzzle whenever a tile is not
selected. When a tile is selected, if it contains a terminus, it will
now show all of the colors in use by the terminus. Masks can now
optionally display a message while they are active.