Fix bug with mask and beam collision
When reloading from state there is an edge case where a beam will get to
a portal and show the mask for user choice on where to go, but then
subsequently gets collided with by another beam. Previously this would
result in a collision and the mask still being displayed, whereas if the
beam entering the portal and causing the mask were the last one to be
turned on in real-time, it would not have displayed the mask at all.

On collision, if there is an open mask it is now checked to see if it is
associated with a beam involved in the collision, and if so, it is
unmasked, which will prevent this bug.

There is also some minor re-organization of the Puzzle.Mask class to
make it more configurable. For example, onUnmask was added, which will
now be called whenever the mask is unmasked.