WIP: refactoring of beam state
This started with an attempt to fix toggling of a terminus that already
has a beam connected to it. It was finally time to tackle the messiness
of beam state, which now has concrete classes for each different state.

Everything is working now except beam merging on portal exit, which is
causing an infinite loop.