Fixing various collision bugs
- fixes bugs relating to reflector collisions involving multiple beams.
  beams will now correctly not collide when on opposite sides of the
reflector. beams will now properly collide if one reflects back at self
and another uses the same side of the reflector.
- fix improper beam path cleanup on removal causing issues with
  collision detection. paths now match the beam visuals.
- fix beams updating too greedily when events are handled. updates now
  happen in the order events happen.

Improved puzzle 005, now it requires using the move modifier, previously
it could be completed using just swap.

Puzzle 007 is a work in progress, and 009 was added for collision
testing and will be removed later.