Bug fixes for beam path, wall item improvements
Tracking of segmentIndex on beam steps has been updated to indicate the
earliest segmentIndex for the step instead of the last. This fixed an
issue where segments were left behind when history is updated. This also
made it easy to clean up empty paths on history updates.

The Wall item UI has been updated so that a wall segment sits flush
against the tile instead of having slanted edges which point towards the
center.

Also fixed a bug with the infinite loop exit logic where it was
triggering unnecessarily due to the last update state not being cleared
out in between updates.