Initial refactor of user interactions
Refactored all interactions to use pointer events instead of mouse and
touch events. Implemented pinch zooming for multi-touch enabled devices
and re-implemented pan support for pointer events.

This is mostly ready to go, but there is some odd behavior when pinch
zooming occasionally when your fingers get close to the edge of the
canvas (it makes the zooming stutter a little bit).