Improve beam collision logic for portals
Handles beams entering/exiting portals in different directions. So long
as they don't try to exit in the same direction, or one tries to enter
at the opposite direction of one exiting, there shouldn't be any
collisions. In other words, beams are free to use any open direction
when entering/exiting a portal.

Step.Stop was also removed in favor of just returning currentStep, which
will cause processing to stop anyways since it will detect the same
point twice in a row.