Improve puzzle 011, fix a few bugs
Updated the puzzle to include a non-directional portal to further
illustrate the rules of portals. The portal mask has been updated to
select the tile the mask applies to before applying the mask (which
ensures the toolbar for that tile is selected). In addition, the
before/after edit actions on tiles have been updated to disable all
modifiers when a mask is active on that tile. The use of modifiers while
a mask was active was causing bugs, especially in the case of a portal.

This also includes a bug fix for updating beam references in a tile.
Previously, all beam references were removed from a tile when a step for
that beam was removed from a tile. However, sometimes there are multiple
steps for a single beam in a tile, so the beam reference only needs to
be removed if there are no longer any steps in that tile for the beam.