Fix bug in beam state related to portals
When there are multiple options for an exit portal, the user must
decide, and the decision gets stored in beam state. That state was only
being cleared when the beam was turned off, but it should be cleared
whenever the step that decision is related to changes (for example, if
the entry portal is rotated).

This also cleans up the storage of these decisions in beam state, moving
them into a portal specific key that the portal item will maintain.