Various fixes related to portals
Items can now specify 'immutable' in the state configuration, which
should basically act the same as setting each individual modifier key to
false (e.g. movable, toggleable, rotatable, etc). Beam and wall items
will have immutable set by default.

Item IDs are now retained in cache, making them stable IDs. Previously,
item IDs were re-generated when the page was refreshed, which meant that
if an item had been moved its ID could change. The generation and
storage of item IDs on the loading of a puzzle is not stored as a delta,
so it won't show up as a move and won't be undo-able.

Portal logic has been reverted back to just checking if a direction is
already occupied. This means beams cannot be merged through an exit
portal. That functionality will be added later as a separate item.
Additionally, portal directions cache is now cleared if a portal is
moved, since the beams connected to that portal will need to re-evaluate
their histories.

Swap now requires tiles to contain movable items, instead of any items.