Add mask.onUpdate
This allows the mask to be re-evaluated when it gets dequeued, with the
potential to cancel the mask. This is because state can change in
between when the mask is queued and when it gets processed. For example,
if multiple portals are reached at the same time, the resolution of one
of the portals may cause a subsequent portal mask to no longer be valid.