Complete initial version of puzzle 012
Still needs test coverage, but I have confirmed that it is solvable as
is. An additional change in this commit is the addition of an "is not
immutable" filter for modifiers when displaying the indicator badge on
tiles. I think the primary purpose of the indicator is to show the user
they can interact with something in that tile, and immutable modifiers
are not interactable, so it just creates noise. The lock and immutable
modifiers mostly apply when moving something anyways.