Improve UI/UX
Make the header and footer floating and centered. This is more ergnomic
on large screens, and not much of a difference on smaller screens.

Instead of using an indicator dot if something in a tile is
interactable, just show all of the modifiers that have been applied to a
tile.

TODO:
 - there will need to be some new rules introduced for tiles that
   contain six modifiers to prevent adding any additional modifiers,
since there won't be any room to display them in the tile. This is not
currently possible as there are only four actionable modifiers plus the
lock at the moment, unless duplicate actions are added.
 - the footer needs to be re-designed. It should always display actions
   which are usable on any tile (e.g. not locked). The message should
always be displayed below the available actions. If there is no message
and no actions available outside of specific tiles, the footer should
not be visible at all.