Begin refactor of modifiers UI
Global messages have been moved to the header. Tiles can now have their
own messages, which will be displayed above actions in the footer.
Immutable modifiers are no longere displayed in the footer, since they
are already displayed within tiles. The actions in the footer are now
separated into locked actions (left side, belong to the selected tile)
and un-locked actions (right side, do not belong to any tile).

Still to do: populating the action bar correctly when a tile or not tile
is selected. When no tile is selected, it should show all un-locked
actions disabled. When a tile is selected, it should show all actions
enabled. When an un-locked action is used on a tile with a lock, that
action should become locked and it will move from the right side to the
left side.