Initial global modifiers integration
Added support for "global" modifiers, or, modifiers that do not belong
to a tile. This is still pretty rough and needs a few more iterations.

This also includes a few fixes for bugs that were observed, as well as
some simplification of puzzle state updating in relation to when
multiple changes are made when a modifier is invoked -- state is now
always saved at the end of invocation right before the beams are
updated.