Add moves to state
This allows tracking moves seperately from deltas, since sometimes
multiple deltas belong to a single move. This also allows IDs for all
items that are stored in state to be generated and persisted without
interfering with moves. This was necessary to ensure that modifiers
always show up in the same place in the toolbar since they are sorted by
their persisted IDs.