Refactor the layout data structure
This is a pretty sizable change, but it makes creating and rendering the
layout programmatically drastically simpler. Instead of using an array of
arrays with null in place of empty tiles, tiles are now stored in an object
by their offset (row, and then column). The offsets are all relative to the
center of the canvas, or 0,0. This allows the data and rendering to both use
the same mechanism to calculate the points where the tiles live.

This change will break all the existing puzzle configurations until they are
converted to the new format. The cache version number will need to be updated
upon completion of this as well, to force any users with stale cache to pull
in the new configuration on page load.