Begin preparation for live puzzle updates
Instead of re-loading state and not tracking changes to state in the
state cache, utilize the same model for state updates that is used
with all other user interactions. This will essentially turn the editor
into a playground where the user can modify the puzzle live and see what
happens, including the ability to undo, redo and reset.