Add steam deployment workflow. Fixes #60 (#61)
Also updated the version check so that it works in the following
way:
- if the version in `package.json` has not changed, simply exit the
release process early (after tests run, before anything else does).
basically this is for cases where something in the repo is changed, but
a new release is not required (for example, updating build
configuration)
- if the version in `package.json` has changed, run the full release
process

The version check can be removed from the `pull-request` workflow. That
workflow will always run the same stuff no matter what (test + make).