Fix gh-pages deployment (#73)
Fixes path to web build, which is now at dist/web. Also allows gh-pages
to be updated regardless of whether or not the build number is being
bumped, since there is no downside to doing it, and it can be useful in
cases like this.

Basically the package version bump is now tied to Steam deployment, and
not to gh-pages deployment.