Add Steamworks SDK support. Fixes #48 (#77)
This will allow logic to bifurcate based on whether the game is running
alongside Steam -- either from Steam itself, or on a machine in which
Steam is running and the user owns the game.

This won't really allow any gating of functionality, as the project is
open source so anyone could simply clone this project and remove any
constraints locally. However, none of the steam integrations will work
without the SDK -- most importantly the future Steam Workshop
implementation.