WIP: editor improvements and optimizations
Editor instances are now cached instead of destroyed/initialized each
time the user selects or de-selects a tile.

The basic structure for handling the editing of different puzzles has
been setup but not implemented. But basically it will involve using
the dropdown and cache to switch between which puzzle is being edited.