Implement a level. Fixes #43 (#80)
A level will basically be a puzzle that incorporates other puzzles
within it. Each referenced puzzle will have an anchor point in the main
puzzle. The referenced puzzles can be conditionally applied in a number
of ways:
- included or excluded based on whether they have been solved by the
user
- tiles included or excluded based on the way the user solved the puzzle

Those are the only conditions necessary for MVP, though the
configuration should probably support more flexible options in the
future (including items, for example).

The completion of this PR should include a simple level that illustrates
that all the functionality works. The actual first level that will be
shipped with the MVP of the game will be developed in a separate PR.